あまりにもコピペだったので共通化した
